+++ frontend/src/App.js
@@ -3,8 +3,8 @@ import { Route, Switch } from "react-router"
 import styled from "styled-components/macro"
 import Header from "./components/header/Header"
 import LoginPage from "./components/pages/LoginPage"
-import Homepage from "./components/pages/Homepage"
-import RecommendationForm from "./components/pages/RecommendationForm"
+import Homepage from "./components/pages/homepage/Homepage"
+import RecommendationForm from "./components/pages/recommandationForm/RecommendationForm"
 import Footer from "./components/footer/Footer"
 
 function App() {
