+++ frontend/src/components/pages/homepage/wine_animation.json
